Got rid of coroutines; added a test executable
Usage of coroutines (as well as the way DrillGeohash was
implemented)
resulted in high memory usage of the index construction operation.
After this change the memory usage dropped by an order of magnitude.

Also added a test executable for building a geohashtree index for
a GeoJSON file using the data found in README.